// packages/ui/src/components/Popover/Popover.type.ts
import type { ReactNode, ReactElement, JSX } from "react";

export enum PopoverCloseSource {
  Click = 0,
  EscapeKeypress = 1,
  FocusOut = 2,
  ScrollOut = 3,
}

export type PopoverPosition = "above" | "below" | "mostSpace" | "cover";
export type PopoverAlignment = "left" | "center" | "right";

export interface PopoverProps {
  /** The content to display inside the popover. */
  children?: ReactNode;
  /** The preferred direction to open the popover. */
  preferredPosition?: PopoverPosition;
  /** The preferred alignment of the popover relative to its activator. */
  preferredAlignment?: PopoverAlignment;
  /** Show or hide the Popover. */
  active: boolean;
  /** The element to activate the Popover. */
  activator: ReactElement;
  /** Use the activator's input element to calculate the Popover position. Defaults to true. */
  preferInputActivator?: boolean;
  /** The element type to wrap the activator with. Defaults to 'div'. */
  activatorWrapper?: string;
  /** Override on the default z-index of 400. */
  zIndexOverride?: number;
  /** Prevents focusing the activator or the next focusable element when the popover is deactivated. */
  preventFocusOnClose?: boolean;
  /** Automatically add wrap content in a section. */
  sectioned?: boolean;
  /** Allow popover to stretch to the full width of its activator. */
  fullWidth?: boolean;
  /** Allow popover to stretch to fit content vertically. */
  fullHeight?: boolean;
  /** Allow popover content to determine the overlay width and height. */
  fluidContent?: boolean;
  /** Remains in a fixed position. */
  fixed?: boolean;
  /** Used to illustrate the type of popover element. */
  ariaHaspopup?:
    | boolean
    | "false"
    | "true"
    | "menu"
    | "dialog"
    | "grid"
    | "listbox"
    | "tree";
  /** Allow the popover overlay to be hidden when printing. */
  hideOnPrint?: boolean;
  /** Callback when popover is closed. */
  onClose: (source: PopoverCloseSource) => void;
  /** The preferred auto focus target defaulting to the popover container. Defaults to 'container'. */
  autofocusTarget?: "none" | "first-node" | "container";
  /** Prevents closing the popover when other overlays are clicked. */
  preventCloseOnChildOverlayClick?: boolean;
  /** Prevents page scrolling when the end of the scrollable Popover overlay content is reached. Defaults to false. */
  captureOverscroll?: boolean;
}